calling an external Airly API

+++ src/main/java/com/agnieszka/piotrowska/weatherCheckApp/model/Arguments.java
@@ -7,7 +7,7 @@ import lombok.*;
 @Data
 public class Arguments {
 
-    public Long installationId;
+    public Integer installationId;
     public Double lat;
     public Double lng;
     public Double maxDistanceKm;
